Don't zip for Mac

+++ build.sh
@@ -141,7 +141,9 @@ else
   fi
 fi
 
-cd $base_dir
-rm -f $base_dir/*.zip
-HASH=$(git rev-parse --short HEAD)
-zip -r m64p-$my_os-$HASH.zip mupen64plus
+if [[ $UNAME != *"Darwin"* ]]; then
+  cd $base_dir
+  rm -f $base_dir/*.zip
+  HASH=$(git rev-parse --short HEAD)
+  zip -r m64p-$my_os-$HASH.zip mupen64plus
+fi
